% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_igv.R
\name{basic2}
\alias{basic2}
\title{abc}
\usage{
basic2(
  table,
  trackName = "demo",
  pval.col = 10,
  chrom.col = 1,
  pos.col = 3,
  ...
)
}
\description{
abc
}
\examples{
data(pax_demo)
# eliminate "uninteresting" loci
pax_demo_filt = pax_demo[pax_demo$P < .1,]
basic2(table=pax_demo_filt)
}
