% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dn8hub.R
\name{dn8_hub}
\alias{dn8_hub}
\title{provisional constructor of a hub to collect all resources and manage connections}
\usage{
dn8_hub()
}
\description{
provisional constructor of a hub to collect all resources and manage connections
}
\note{
The structure is a flat list with components connections, tables (just
a list of tablenames per connection), and status (defaults to "open")
only for use at channing
}
