% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_igv.R
\name{filter_dn8like}
\alias{filter_dn8like}
\title{filter a table with dn8like columns}
\usage{
filter_dn8like(dn8like, chr, start, end, posvbl)
}
\arguments{
\item{dn8like}{data.frame}

\item{chr}{character(1)}

\item{start}{numeric(1)}

\item{end}{numeric(1)}

\item{posvbl}{character(1) column named for position}
}
\description{
filter a table with dn8like columns
}
\examples{
data("gtex_b38_lung_chr20_exc", package="tnt4dn8")
filter_dn8like( gtex_b38_lung_chr20_exc, 20, 60000, 62000, "pos_b38")
}
