% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dn8hub.R
\name{filter_sym}
\alias{filter_sym}
\title{filter a symbol from a table meeting dn8 naming conditions}
\usage{
filter_sym(.data, sym, radius = 0)
}
\arguments{
\item{.data}{a tbl}

\item{sym}{a character(1) gene symbol}

\item{radius}{numeric(1) include `radius` bp upstream and downstream of gene body limits}
}
\description{
filter a symbol from a table meeting dn8 naming conditions
}
\note{
if `pos_b38` is found in the dn8 table, gene addresses from EnsDb.Hsapiens.v79 are used to get gene
addresses in hg38 coordinates.  Otherwise, EnsDb.Hsapiens.v75 is used for gene coordinates.
}
\examples{
if (interactive()) {
 h = dn8_hub()
 chk = h |> tbl("leic_dn8") |> filter_sym("HHIP") |> as.data.frame()
 print(names(chk))
 print(dim(chk))
 print(summary(-log10(chk$p)))
 chk2 = h |> tbl("leic_dn8") |> filter_sym("HHIP", radius=1e6) |> as.data.frame()
 print(dim(chk2))
 print(summary(-log10(chk2$p)))
 print(tables(h))
 lapply(tables(h), function(x) h |> tbl(x))
 dn8_hub_close(h)
}
}
