% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes_ensv75}
\alias{genes_ensv75}
\title{a data frame with gene addresses etc. for hg19 coordinates}
\format{
data.frame
}
\usage{
genes_ensv75
}
\description{
a data frame with gene addresses etc. for hg19 coordinates
}
\keyword{datasets}
