% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes_ensv79}
\alias{genes_ensv79}
\title{a data frame with gene addresses etc. for hg38 coordinates}
\format{
data.frame
}
\usage{
genes_ensv79
}
\description{
a data frame with gene addresses etc. for hg38 coordinates
}
\keyword{datasets}
