% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genetrack_hg19}
\alias{genetrack_hg19}
\title{TnT genetrack from Ensembl v75}
\format{
TnT genetrack
}
\usage{
genetrack_hg19
}
\description{
TnT genetrack from Ensembl v75
}
\keyword{datasets}
