% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtex_b38_lung_chr20_exc}
\alias{gtex_b38_lung_chr20_exc}
\title{small dn8-formatted GTEx excerpt, chr20}
\format{
data.frame
}
\usage{
gtex_b38_lung_chr20_exc
}
\description{
small dn8-formatted GTEx excerpt, chr20
}
\keyword{datasets}
