% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38ToHg19}
\alias{hg38ToHg19}
\title{a chain lifting from hg38 to hg19}
\format{
rtracklayer Chain instance
}
\usage{
hg38ToHg19
}
\description{
a chain lifting from hg38 to hg19
}
\keyword{datasets}
