% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{limgwcat_b38}
\alias{limgwcat_b38}
\title{small dn8-similar GWAS catalog excerpt, chr20}
\format{
data.frame
}
\usage{
limgwcat_b38
}
\description{
small dn8-similar GWAS catalog excerpt, chr20
}
\keyword{datasets}
