% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_igv.R
\name{sym2region}
\alias{sym2region}
\title{produce a list of chrom, start, end, given gene symbol}
\usage{
sym2region(sym = "SOX12", build = "hg38")
}
\arguments{
\item{sym}{character(1)}

\item{build}{character(1)}
}
\description{
produce a list of chrom, start, end, given gene symbol
}
\examples{
sym2region()
}
