% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dn8hub.R
\name{tables}
\alias{tables}
\title{list the 'tables' available in a hub (really the database connection names)}
\usage{
tables(x)
}
\arguments{
\item{x}{instance of dn8_hub}
}
\description{
list the 'tables' available in a hub (really the database connection names)
}
