% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dn8hub.R
\name{tbl.dn8_hub}
\alias{tbl.dn8_hub}
\title{select a tbl from a hub}
\usage{
\method{tbl}{dn8_hub}(src, ...)
}
\arguments{
\item{src}{a dn8_hub instance}

\item{\dots}{should be character(1)}
}
\description{
select a tbl from a hub
}
