% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{tntplot}
\alias{tntplot}
\title{TnT approach}
\usage{
tntplot(tab, snpcolor = "lightblue", genecolor = "gold", gt = NULL)
}
\arguments{
\item{tab}{table}

\item{snpcolor}{character(1)}

\item{genecolor}{character(1)}

\item{gt}{defaults to NULL, otherwise a GeneTrackFromTxDb-like object from TnT}
}
\description{
TnT approach
}
\examples{
if (requireNamespace("TnT") & requireNamespace("TxDb.Hsapiens.UCSC.hg19.knownGene")) {
data(gtex_b38_lung_chr20_exc)
cands = avail_syms_gtex()
cands
chk2 = gtex_b38_lung_chr20_exc |> filter_sym(cands[3], radius=5e5) |> as.data.frame()
print(tntplot(chk2))
}
}
