% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{tntplot2}
\alias{tntplot2}
\title{detailed plot}
\usage{
tntplot2(
  tab,
  snpcolor = "lightblue",
  genecolor = "darkgreen",
  txcolor = "darkred",
  gt = NULL,
  maxp = 0.1
)
}
\arguments{
\item{tab}{data.frame}

\item{snpcolor}{character(1)}

\item{genecolor}{character(1)}

\item{txcolor}{character(1)}

\item{gt}{defaults to NULL, otherwise a GeneTrackFromTxDb-like object from TnT}

\item{maxp}{numeric(1) if non-NULL loci with p-values greater than this are excluded}
}
\description{
detailed plot
}
\examples{
if (requireNamespace("TnT") & requireNamespace("TxDb.Hsapiens.UCSC.hg19.knownGene")) {
data(gtex_b38_lung_chr20_exc)
cands = avail_syms_gtex()
cands
chk2 = gtex_b38_lung_chr20_exc |> filter_sym(cands[3], radius=5e4) |> as.data.frame()
print(tntplot2(chk2))
}
}
