% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{tntplot3}
\alias{tntplot3}
\title{detailed plot with 2 data tracks}
\usage{
tntplot3(
  tab,
  tab2,
  snpcolor = "lightblue",
  genecolor = "darkgreen",
  txcolor = "darkred",
  GT = NULL,
  tt = NULL,
  maxp = 0.1,
  trunc_mlp = 90,
  lab1 = "track1",
  lab2 = "track2"
)
}
\arguments{
\item{tab}{data.frame, main GWAS}

\item{tab2}{data.frame, second set of locations and scores}

\item{snpcolor}{character(1)}

\item{genecolor}{character(1)}

\item{txcolor}{character(1)}

\item{GT}{defaults to NULL, otherwise a GeneTrackFromTxDb-like object from TnT}

\item{tt}{defaults to NULL, otherwise a TxTrackFromTxDb-like object from TnT}

\item{maxp}{numeric(1) if non-NULL loci with p-values greater than this are excluded}

\item{trunc_mlp}{numeric(1) defaults to 300, -log10p greater than this are reset to this value}

\item{lab1}{character(1) label for first scored track}

\item{lab2}{character(1) label for second scored track}
}
\description{
detailed plot with 2 data tracks
}
\examples{
if (requireNamespace("TnT") & requireNamespace("TxDb.Hsapiens.UCSC.hg38.knownGene")) {
data(gtex_b38_lung_chr20_exc)
data(limgwcat_b38)
cands = avail_syms_gtex()
cands
chk2 = gtex_b38_lung_chr20_exc |> filter_sym(cands[3], radius=5e4) |> as.data.frame()
chk3 = limgwcat_b38 |> filter_sym(cands[3], radius=5e4) |> as.data.frame()
print(tntplot3(chk2, chk3, lab1="GTEx Lung", lab2="EBI GWAS cat"))
}
}
