% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vjcdemo.R
\name{vjcdemo}
\alias{vjcdemo}
\title{test single GWAS track}
\usage{
vjcdemo(
  table,
  name = "gwasdemo",
  chrom.col = 1,
  pos.col = 3,
  pval.col = 10,
  ...
)
}
\arguments{
\item{table}{data.frame}

\item{name}{character(1) used as trackName}

\item{chrom.col}{numeric(1)}

\item{pos.col}{numeric(1)}

\item{pval.col}{numeric(1)}

\item{\dots}{passed to GWASTrack}
}
\description{
test single GWAS track
}
\examples{
data(pax_demo)
vjcdemo(pax_demo)
}
