% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{tntplot}
\alias{tntplot}
\title{TnT approach}
\usage{
tntplot(tab, snpcolor = "lightblue", genecolor = "gold", gt = NULL)
}
\arguments{
\item{tab}{table}

\item{snpcolor}{character(1)}

\item{genecolor}{character(1)}

\item{gt}{defaults to NULL, otherwise a GeneTrackFromTxDb-like object from TnT}
}
\description{
TnT approach
}
\examples{
if (requireNamespace("TnT") & requireNamespace("TxDb.Hsapiens.UCSC.hg19.knownGene")) {
h = dn8_hub()
chk2 = h |> tbl("leic_dn8") |> filter_sym("HHIP", radius=1e6) |> as.data.frame()
dn8_hub_close(h)
print(tntplot(chk2))
}
}
